requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pytest
httpx
python-dotenv
alembic
pydantic-settings